rewrite filter function
rewrite the switch statement to check member_name instead of message_type, this
allow to mege the CloseNotification and NotificationClosed cases. the case of
the notify call beeing successfull or not is also merged.

the monitor signals are now emitted before the next Gtk redraw cycle, making
sure that the tray got updated in the current cycle.

Signed-off-by: Gustavo Marques <pushstarttocontinue@outlook.com>